requirements.txt
praw
nltk
tqdm
python-dotenv
requests
beautifulsoup4
pandas
matplotlib
seaborn
wordcloud